% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_USSL.R
\name{calculate_sar}
\alias{calculate_sar}
\title{Calculate SAR}
\usage{
calculate_sar(df, convert_units = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{convert_units}{logical, for conversion to meq/l}
}
\value{
SAR values
}
\description{
This function calculates the Sodium Adsorption Ratio (SAR) for water quality.
}
