% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Wilcox.R
\docType{data}
\name{CC}
\alias{CC}
\title{Color palette for Wilcox diagram}
\format{
An object of class \code{character} of length 12.
}
\usage{
CC
}
\description{
This vector defines the color palette used in the Wilcox diagram.
It contains a sequence of color names.
}
\keyword{data}
