% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrigationALL.R
\name{irrigationALL}
\alias{irrigationALL}
\title{Irrigation Water Quality Index Calculations}
\usage{
irrigationALL(df, convert_to_meq = TRUE)
}
\arguments{
\item{df}{Dataframe containing necessary variables.}

\item{convert_to_meq}{Logical, indicating whether to convert concentrations to meq/L (default: TRUE).}
}
\value{
Dataframe containing calculated indices: SAR, MAR, SCAR, RSC, RSBC, PI, KR, NaPercentage, and PS.
}
\description{
This function calculates multiple water quality indices for irrigation.
}
