% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KR.R
\name{KR}
\alias{KR}
\title{Kelly Ratio (KR) Calculation}
\usage{
KR(df, convert_to_meq = TRUE)
}
\arguments{
\item{df}{Dataframe containing necessary variables (Na, Ca).}

\item{convert_to_meq}{Logical, indicating whether to convert concentrations to meq/L (default: TRUE).}
}
\value{
KR value.
}
\description{
Kelly Ratio (KR) Calculation
}
