% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAR.R
\name{MAR}
\alias{MAR}
\title{Magnesium Adsorption Ratio (MAR) Calculation}
\usage{
MAR(df, convert_to_meq = TRUE)
}
\arguments{
\item{df}{Dataframe containing necessary variables (Mg, Ca).}

\item{convert_to_meq}{Logical, indicating whether to convert concentrations to meq/L (default: TRUE).}
}
\value{
MAR value.
}
\description{
Magnesium Adsorption Ratio (MAR) Calculation
}
