% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Na%.R
\name{NaPercentage}
\alias{NaPercentage}
\title{Sodium Percentage (NaPercentage) Calculation}
\usage{
NaPercentage(df, convert_to_meq = TRUE)
}
\arguments{
\item{df}{Dataframe containing necessary variables (Na, Ca, Mg, K).}

\item{convert_to_meq}{Logical, indicating whether to convert concentrations to meq/L (default: TRUE).}
}
\value{
Sodium Percentage value.
}
\description{
Sodium Percentage (NaPercentage) Calculation
}
