% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI.R
\name{PI}
\alias{PI}
\title{Permeability Index (PI) Calculation}
\usage{
PI(df, convert_to_meq = TRUE)
}
\arguments{
\item{df}{Dataframe containing necessary variables (Na, HCO3, Ca, Mg).}

\item{convert_to_meq}{Logical, indicating whether to convert concentrations to meq/L (default: TRUE).}
}
\value{
PI value.
}
\description{
Permeability Index (PI) Calculation
}
