% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PS.R
\name{PS}
\alias{PS}
\title{Potential Salinity (PS) Calculation}
\usage{
PS(df, convert_to_meq = TRUE)
}
\arguments{
\item{df}{Dataframe containing necessary variables (Cl, SO4).}

\item{convert_to_meq}{Logical, indicating whether to convert concentrations to meq/L (default: TRUE).}
}
\value{
PS value.
}
\description{
Potential Salinity (PS) Calculation
}
