% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSBC.R
\name{RSBC}
\alias{RSBC}
\title{Residual Sodium Bicarbonate (RSBC) Calculation}
\usage{
RSBC(df, convert_to_meq = TRUE)
}
\arguments{
\item{df}{Dataframe containing necessary variables (HCO3, Ca).}

\item{convert_to_meq}{Logical, indicating whether to convert concentrations to meq/L (default: TRUE).}
}
\value{
RSBC value.
}
\description{
Residual Sodium Bicarbonate (RSBC) Calculation
}
