% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSC.R
\name{RSC}
\alias{RSC}
\title{Residual Sodium Carbonate (RSC) Calculation}
\usage{
RSC(df, convert_to_meq = TRUE)
}
\arguments{
\item{df}{Dataframe containing necessary variables (HCO3, CO3, Ca, Mg).}

\item{convert_to_meq}{Logical, indicating whether to convert concentrations to meq/L (default: TRUE).}
}
\value{
RSC value.
}
\description{
Residual Sodium Carbonate (RSC) Calculation
}
