% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAR.R
\name{SAR}
\alias{SAR}
\title{Sodium Adsorption Ratio (SAR) Calculation}
\usage{
SAR(df, convert_to_meq = TRUE)
}
\arguments{
\item{df}{Dataframe containing necessary variables (Na, Ca, Mg, K).}

\item{convert_to_meq}{Logical, indicating whether to convert concentrations to meq/L (default: TRUE).}
}
\value{
SAR value.
}
\description{
Sodium Adsorption Ratio (SAR) Calculation
}
