% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAR.R
\name{SCAR}
\alias{SCAR}
\title{Sodium Adsorption Ratio (SCAR) Calculation}
\usage{
SCAR(df, convert_to_meq = TRUE)
}
\arguments{
\item{df}{Dataframe containing necessary variables (Na, Ca).}

\item{convert_to_meq}{Logical, indicating whether to convert concentrations to meq/L (default: TRUE).}
}
\value{
SCAR value.
}
\description{
Sodium Adsorption Ratio (SCAR) Calculation
}
